Spatie Exercises

+++ WebSecService/app/Http/Controllers/Web/ProductsController.php
@@ -86,6 +86,10 @@ class ProductsController extends Controller
 
     public function delete(Request $request, Product $product) 
     {
+        if (!auth()->user()->hasPermissionTo('delete_products')) {
+            abort(401);
+        }
+
         $product->delete();
         return redirect()->route('products_list');
     }
